Start reorganizing steps for Noruba
