+++ noruba/features/lib/step_definitions/filesystem.rb
@@ -1,0 +1,4 @@
+def write_file(path, content)
+  FileUtils.mkdir_p(File.dirname(path))
+  File.open(path, 'w') { |file| file.write(content) }
+end
